web: Add loadTest
The load test floods the application with multiple reservation requests
in parallel. This should stress all parts of the application because the
web component has to handle the request, validate it and pass it on to
the georesolver, which has to validate it, try to find a suitable slot
write the result to the database of any and inform the newsbeeper. The
newsbeeper has to validate the request and store the resulting
notifications in the database.

This generates load on all components including the database (data
retrival and storage) and the MoM (sending of messages between
components).